Checkin page components
