+++ src/containers/Patient/checkIn.tsx
@@ -1,406 +1,17 @@
-import React, { Children, useState } from "react";
+import React, { useState } from "react";
 import { NextPage, NextPageContext } from "next";
 import Modal from "components/Modal/modal";
 import Button from "components/Button/button";
 import PatientDetails from "components/CheckIn/PatientDetails";
-import CheckinItem from "components/CheckIn/CheckinItem";
 import PhysicalExaminationModal from "components/CheckIn/PhysicalExaminationModal";
+import MedicalRecordsItems from "components/CheckIn/MedicalRecordsItems";
+import CheckinItemsDisplay from "components/CheckIn/CheckinItemsDisplay";
+import PhysicalCheckResult, {
+  IphysicalExamination,
+} from "components/CheckIn/PhysicalCheckResult";
 
-const MedicalRecordsItems = (props: {
-  onRecordItemTypeUpdate: Function;
-  children: React.FunctionComponentFactory;
-}) => {
-  const handleAddMedicalRecordItem = (item: string) => {
-    console.log(item);
-    props.onRecordItemTypeUpdate(item);
-  };
-
-  return (
-    <>
-      <div className="medical__records">
-        <div className="medical__records--buttons">
-          <div className="buttons__list">
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Chief complain")}
-            >
-              + Chief complain
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Physical Examination")}
-            >
-              + Physical Examination
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Clinical Signs")}
-            >
-              + Clinical Signs
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Tentative Diagnosis")}
-            >
-              + Tentative Diagnosis
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Diagnosis Test")}
-            >
-              + Diagnosis Test
-            </Button>
-          </div>
-
-          <div className="buttons__list">
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Final Diagnosis")}
-            >
-              + Final Diagnosis
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Treatment")}
-            >
-              + Treatment
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Vaccination")}
-            >
-              + Vaccination
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Note")}
-            >
-              + Note
-            </Button>
-            <Button
-              type="primary"
-              onClick={() => handleAddMedicalRecordItem("Medical Bill")}
-            >
-              + Medical Bill
-            </Button>
-          </div>
-        </div>
-        {props.children}
-      </div>
-    </>
-  );
-};
-
-const CheckinItemsDisplay = (props: { children: any }) => {
-  return (
-    <div className="checkin__items">
-      <div>
-        <form className="checkin-form">
-          <div className="form-item">
-            <label>Check In</label>
-
-            <input type="text" disabled defaultValue={new Date().toString()} />
-          </div>
-
-          <div className="form-item">
-            <label>Visit Type</label>
-
-            <select disabled>
-              {[{ value: "Follow Up", label: "Follow Up" }].map((opt) => {
-                return <option value={opt.value}>{opt.label}</option>;
-              })}
-            </select>
-          </div>
-        </form>
-      </div>
-      {props.children}
-    </div>
-  );
-};
-
-interface IphysicalExamination {
-  respiratoryRate: string;
-  rectalTemperature: string;
-  ocularMucousMembrane: string;
-  oralMucousMembrane: string;
-  nasalCavity: string;
-  integumentFur: string;
-  natureOfLesion: string;
-  locationOfLesion: string;
-  ectoparasite: string;
-  generalDisposation: string;
-  conformation: string;
-  anyDiarrhea: string;
-  natureOfDiarrhea: string;
-  lungsSound: string;
-  anyLameness: string;
-  lamenessLocation: string;
-  rectalExamination: string;
-  prepuceVulvaExamination: string;
-}
-
-const PhysicalCheckResult = (props: {
-  physicalExaminationResult: IphysicalExamination;
-  showModal: Function;
-}) => {
-  const { physicalExaminationResult, showModal } = props;
-  return (
-    <CheckinItem
-      title="Physical Examination"
-      actions={
-        <>
-          <input defaultValue={new Date().toString()} />
-          <Button onClick={() => showModal(true)}>Add New Results</Button>
-          <Button onClick={() => {}}>Delete Results</Button>
-        </>
-      }
-    >
-      <div className="patient__checkin__exam__result">
-        {physicalExaminationResult.respiratoryRate && (
-          <ul className="exam__result__list">
-            {[
-              {
-                label: "Rectal Temperature",
-                value: physicalExaminationResult.rectalTemperature,
-              },
-              {
-                label: "Respiratory Rate",
-                value: physicalExaminationResult.respiratoryRate,
-              },
-              {
-                label: "Ocular Mucous Membrane",
-                value: physicalExaminationResult.ocularMucousMembrane,
-              },
-              {
-                label: "Oral Mucous Membrane",
-                value: physicalExaminationResult.oralMucousMembrane,
-              },
-              {
-                label: "Nasal cavity",
-                value: physicalExaminationResult.nasalCavity,
-              },
-              {
-                label: "Integument/Fur",
-                value: physicalExaminationResult.integumentFur,
-              },
-              {
-                label: "Nature of Lesion ",
-                value: physicalExaminationResult.natureOfLesion,
-              },
-              {
-                label: "Location of Lesion ",
-                value: physicalExaminationResult.locationOfLesion,
-              },
-              {
-                label: "Ectoparasite",
-                value: physicalExaminationResult.ectoparasite,
-              },
-              {
-                label: "General Disposation",
-                value: physicalExaminationResult.generalDisposation,
-              },
-              {
-                label: "Conformation",
-                value: physicalExaminationResult.conformation,
-              },
-              {
-                label: "Any Diarrhea",
-                value: physicalExaminationResult.anyDiarrhea,
-              },
-              {
-                label: "Nature of Diarrhea",
-                value: physicalExaminationResult.natureOfDiarrhea,
-              },
-              {
-                label: "Lungs Sound",
-                value: physicalExaminationResult.lungsSound,
-              },
-              {
-                label: "Any Lameness",
-                value: physicalExaminationResult.anyLameness,
-              },
-              {
-                label: "Lameness Location",
-                value: physicalExaminationResult.lamenessLocation,
-              },
-              {
-                label: "Rectal examination",
-                value: physicalExaminationResult.rectalExamination,
-              },
-              {
-                label: "Prepuce/Vulva Examination",
-                value: physicalExaminationResult.prepuceVulvaExamination,
-              },
-            ].map(({ label, value }) => {
-              return (
-                <li className="exam__result__item">
-                  <span className="exam__result__item--name">{label}</span>
-                  <span className="exam__result__item--value">{value}</span>
-                </li>
-              );
-            })}
-          </ul>
-        )}
-      </div>
-    </CheckinItem>
-  );
-};
-
-const CheckedinItemsDisplay = (props: {
-  onActiveItemChange: Function;
-  children: any;
-  activeNavItem: string;
-}) => {
-  const handleChangeActiveItem = (newItem: string) => {
-    props.onActiveItemChange(newItem);
-  };
-
-  return (
-    <div className="checkedin__items">
-      <ul className="checkedin__navbar">
-        {[
-          "Client Details",
-          "Patient Details",
-          "Medical Records",
-          "Vaccination",
-          "Laboratory",
-          "Radiology",
-          "Appointment",
-        ].map((name: string) => {
-          return (
-            <li
-              className={`checkedin__navbar--item${
-                name === props.activeNavItem ? " active--item" : ""
-              }`}
-              onClick={() => handleChangeActiveItem(name)}
-            >
-              {name}
-            </li>
-          );
-        })}
-      </ul>
-      <div className="checkedin__items--content">{props.children}</div>
-    </div>
-  );
-};
-
-const MedicalReportModalContentTemplate = (props: {
-  children: any;
-  title: string;
-  onAdd: Function;
-  onCancel: Function;
-}) => {
-  return (
-    <div className="medical__report__template">
-      <div className="medical__report__template--head">
-        <span className="template__head--title">{props.title}</span>
-        <span className="template__head--date">
-          <span>Date Recorded</span>{" "}
-          <input type="text" disabled defaultValue={new Date().toString()} />
-        </span>
-      </div>
-
-      <div className="medical__report__template--content">{props.children}</div>
-
-      <div className="medical__report__template--footer">
-        <Button type="primary" onClick={(e) => props.onAdd(e)}>
-          Add
-        </Button>{" "}
-        <Button onClick={() => props.onCancel()}>Cancel</Button>
-      </div>
-    </div>
-  );
-};
-
-const ChiefComplainReport = (props: {
-  title: string;
-  onAdd: Function;
-  onCancel: Function;
-}) => {
-  const handleGetReport = (e: Event) => {
-    e.preventDefault();
-    props.onAdd(formValues);
-  };
-  const [formValues, setFormValues] = useState<{ chiefComplain: string }>({
-    chiefComplain: "",
-  });
-
-  const handleInputChange = (event: {
-    persist: () => void;
-    target: { name: any; value: any };
-  }) => {
-    event.persist();
-    setFormValues((formValues: any) => ({
-      ...formValues,
-      [event.target.name]: event.target.value,
-    }));
-  };
-
-  return (
-    <MedicalReportModalContentTemplate
-      title={props.title}
-      onAdd={handleGetReport}
-      onCancel={props.onCancel}
-    >
-      <form className="medical__report__form">
-        <div className="medical__report__form--input">
-          <label>Rectal Temperatur (°C)</label>
-          <textarea
-            name={"chiefComplain"}
-            rows="10"
-            onChange={handleInputChange}
-          >
-            {formValues.chiefComplain}
-          </textarea>
-        </div>
-      </form>
-    </MedicalReportModalContentTemplate>
-  );
-};
-
-const MedicalRecordModal = ({
-  show,
-  currentModal,
-  getResult,
-  closeModal,
-}: {
-  show: boolean;
-  currentModal: string;
-  getResult: Function;
-  closeModal: Function;
-}) => {
-  console.log(currentModal);
-  return (
-    <Modal
-      fullMode={true}
-      visible={show}
-      closeModal={() => {
-        closeModal(false);
-      }}
-    >
-      {currentModal === "Chief complain" && (
-        <ChiefComplainReport
-          onCancel={closeModal}
-          title={currentModal}
-          onAdd={(data: {}) => getResult(data, currentModal)}
-        />
-      )}
-      {currentModal === "Physical Examination" && (
-        <>Toggle Physical Examination</>
-      )}
-      {currentModal === "Clinical Signs" && <>Toggle Clinical Signs</>}
-      {currentModal === "Tentative Diagnosis" && (
-        <>Toggle Tentative Diagnosis</>
-      )}
-      {currentModal === "Diagnosis Test" && <>Toggle Diagnosis Test</>}
-      {currentModal === "Final Diagnosis" && <>Toggle Final Diagnosis</>}
-      {currentModal === "Treatment" && <>Toggle Treatment</>}
-      {currentModal === "Vaccination" && <>Toggle Vaccination</>}
-      {currentModal === "Note" && <>Toggle Note</>}
-      {currentModal === "Medical Bill" && <>Toggle Medical Bill</>}
-    </Modal>
-  );
-};
+import CheckedinItemsDisplay from "components/CheckIn/CheckedinItemsDisplay";
+import MedicalRecordModal from "components/CheckIn/MedicalRecordModal";
 
 const PatientCheckIn: NextPage<{ patientId: string }> = ({ patientId }) => {
   const [physicalExaminationResult, setPhysicalExaminationResult] = useState<
@@ -569,7 +180,7 @@ const PatientCheckIn: NextPage<{ patientId: string }> = ({ patientId }) => {
           setShowMedicalModal(false);
           setMedicalContentState("");
         }}
-        getResult={(data, field) => {
+        getResult={(data: object, field: string) => {
           handleGetMedicalReportData(data, field);
         }}
         currentModal={medicalContentState}
